# Conflicts:
#	proto/common.proto
#	src/batch/src/executor/mod.rs
#	src/batch/src/executor/row_seq_scan.rs
#	src/batch/src/rpc/service/task_service.rs
#	src/batch/src/task/task_manager.rs
#	src/common/src/session_config/mod.rs
#	src/compute/tests/integration_tests.rs
#	src/frontend/src/handler/flush.rs
#	src/frontend/src/scheduler/distributed/query.rs
#	src/frontend/src/scheduler/hummock_snapshot_manager.rs
#	src/frontend/src/scheduler/local.rs
#	src/rpc_client/src/compute_client.rs
#	src/storage/hummock_sdk/src/lib.rs

+++ src/frontend/src/optimizer/plan_node/batch_limit.rs
@@ -83,8 +83,8 @@ impl ToDistributedBatch for BatchLimit {
 impl ToBatchProst for BatchLimit {
     fn to_batch_prost_body(&self) -> NodeBody {
         NodeBody::Limit(LimitNode {
-            limit: self.logical.limit() as u64,
-            offset: self.logical.offset() as u64,
+            limit: self.logical.limit(),
+            offset: self.logical.offset(),
         })
     }
 }
